requirements.txt
ipython
django<1.10
celery<=3.2
django-debug-toolbar==1.6
redis